Added stop id and some changes
Changed where teh data was pulling from, and added the stop id for some additional information.  Also changed to pandas dataframe, sampling instead of the ptyroch randomsampling, to keep it in more or less sequential time based order, by day and scheduled time for the stop.  So the lstm can pick up on the time dependencies and it seems to have done better.  JUst cant pick up too much on heavy outliers with big delays.